OrderProduct sizes are in decimal.

+++ db/migrate/20231204010016_create_order_products.rb
@@ -5,7 +5,7 @@ class CreateOrderProducts < ActiveRecord::Migration[6.1]
       t.integer :user_id
       t.integer :order_id
       t.integer :quantity
-      t.integer :size
+      t.float :size
       t.boolean :in_cart
       t.timestamps
     end
